requirements.txt
aws-cdk-lib==2.58.1
constructs>=10.0.0,<11.0.0

# pip install boto3
boto3==1.26.62
botocore==1.29.62

# pip install elasticsearch
elasticsearch>=7.0.0,<7.11
opensearch-py==2.0.1
requests==2.23.0
requests-aws4auth==0.9
